Update polygon class

+++ src/com/bemy/protoMap/core/geom/Line.java
@@ -2,9 +2,9 @@ package com.bemy.protoMap.core.geom;
 
 public class Line extends shape{
 	private static final long serialVersionUID = 1110000L;
-	private int startX,endX,startY,endY;
+	private float startX,endX,startY,endY;
 	
-	public Line(int p_startX,int p_endX, int p_startY,int p_endY){
+	public Line(float p_startX,float p_endX, float p_startY,float p_endY){
 		//this.startPoint=new Vector2f( startX, startY);
 		//this.endPoint=new Vector2f( endX, endY);
 		startX=p_startX;
@@ -15,12 +15,12 @@ public class Line extends shape{
 	}
 	@Override
 	public void setPixels() {
-		int dx=Math.abs(endX-startX);
-		int dy=Math.abs(endY-startY);
+		float dx=Math.abs(endX-startX);
+		float dy=Math.abs(endY-startY);
 		float slope=dy/dx;
-		int dp=2*dy-dx;
-		int sx=(endX-startX)>0?1:-1;
-		int sy=(endY-startY)>0?1:-1;
+		float dp=2*dy-dx;
+		float sx=(endX-startX)>0?1:-1;
+		float sy=(endY-startY)>0?1:-1;
 		
 		this.pixels.add(new Vector2f(0, 0));
 		if( slope < 1){
